refactor(characters): extract Characters.Preparation
Final extraction slice of #134. Preparation state, activate, and
add_to_typed_inventory move with all their helpers — the shared
prep-context pipeline from #126 keeps the read and write paths in one
module. The Characters facade is now pure delegation; its moduledoc
documents the submodule-per-concern layout.
